azure-functions/summarization-pipeline/requirements.txt
azure-functions
azure-storage-blob
azure-identity
anthropic>=0.23.0
typing-extensions
